backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
boto3==1.40.1
botocore==1.40.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
litellm==1.74.14
livekit-api==1.0.3
livekit-protocol==1.0.4
lxml==6.0.0
mako==1.3.10
markupsafe==3.0.2
multidict==6.6.3
openai==1.98.0
packaging==25.0
passlib==1.7.4
postgrest==1.1.1
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyjwt==2.10.1
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
pyyaml==6.0.2
realtime==2.6.0
redis==6.2.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.42
starlette==0.47.2
storage3==0.12.0
strenum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.4
tqdm==4.67.1
types-protobuf==6.30.2.20250703
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
